[REFACTOR(MainActivity)] 각 fragment 미리 생성 및 enum class 추가
