+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
             android:exported="false"
             android:screenOrientation="sensor" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
